+++ @navikt/core/react/src/date/datepicker/index.ts
@@ -1,0 +1,12 @@
+export { default as DatePicker, type DatePickerProps } from "./DatePicker";
+export {
+  useDatepicker,
+  useRangeDatepicker,
+  type DateValidationT,
+  type RangeValidationT,
+} from "../hooks";
+export {
+  DatePickerStandalone,
+  type DatePickerStandaloneProps,
+} from "./DatePickerStandalone";
+export { DatePickerInput, type DateInputProps } from "../parts/DateInput";
