--- @navikt/core/react/src/date/index.ts
@@ -1,17 +1,0 @@
-export {
-  default as DatePicker,
-  type DatePickerProps,
-} from "./datepicker/DatePicker";
-export { type DatePickerStandaloneProps } from "./datepicker/DatePickerStandalone";
-export {
-  useDatepicker,
-  useMonthpicker,
-  useRangeDatepicker,
-  type DateValidationT,
-  type MonthValidationT,
-  type RangeValidationT,
-} from "./hooks";
-export { default as MonthPicker } from "./monthpicker/MonthPicker";
-export { type MonthPickerStandaloneProps } from "./monthpicker/MonthPickerStandalone";
-export { type MonthPickerProps } from "./monthpicker/types";
-export { type DateInputProps } from "./parts/DateInput";
