+++ @navikt/core/react/src/date/monthpicker/index.ts
@@ -1,0 +1,9 @@
+"use client";
+export { useMonthpicker, type MonthValidationT } from "../hooks";
+export { default as MonthPicker } from "./MonthPicker";
+export { type MonthPickerProps } from "./types";
+export {
+  MonthPickerStandalone,
+  type MonthPickerStandaloneProps,
+} from "./MonthPickerStandalone";
+export { MonthPickerInput } from "../parts/DateInput";
