+++ @navikt/core/react/src/index.ts
@@ -5,7 +5,23 @@ export { Button, type ButtonProps } from "./button";
 export { Chat, type ChatProps } from "./chat";
 export { Chips, type ChipsProps } from "./chips";
 export { CopyButton, type CopyButtonProps } from "./copybutton";
-export * from "./date";
+export {
+  DatePicker,
+  useDatepicker,
+  useRangeDatepicker,
+  type DatePickerProps,
+  type DatePickerStandaloneProps,
+  type DateInputProps,
+  type DateValidationT,
+  type RangeValidationT,
+} from "./date/datepicker";
+export {
+  MonthPicker,
+  useMonthpicker,
+  type MonthPickerProps,
+  type MonthPickerStandaloneProps,
+  type MonthValidationT,
+} from "./date/monthpicker";
 export * from "./dropdown";
 export { ExpansionCard, type ExpansionCardProps } from "./expansion-card";
 export * from "./form";
